Add files via upload
I had originally created individual functions for each button press like the infamous isEven experiment but saw how it was just really inefficient. I tried to make a map, storing the audio as the value and the id as the key pair but I was struggling with Object Reference Errors. Taking a break from going through all the documentation, I decided to instead start with a switch statement and it worked pretty well.