Add files via upload
Ok a couple things. The big change was making it more accessible for mobile devices. No more runoff or overflow. I changed the styling of the buttons so that they would more closely resemble my own guitar, and I changed their position so that they would resemble an actual fret. I think all I really have left (if I'm not going to update the backgrounds because it looks low budget) is replacing the placeholder text.